requirements.txt
aniso8601==5.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
flask-marshmallow==0.9.0
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
marshmallow==2.18.1
python-dateutil==2.8.0
pytz==2018.9
relativetimebuilder==0.2.0
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.0
urllib3==1.24.1
Werkzeug==0.14.1
pywebpush